+++ modules/Users/includes/General_Info.inc.php
@@ -173,12 +173,12 @@ echo '</td></tr><tr class="st"><td colspan="2">';
 
 echo NoInput( makeLogin( issetVal( $staff['LAST_LOGIN'], '' ) ), _( 'Last Login' ) );
 
-echo '</td></tr></table><hr />';
-
-echo '<table class="width-100p valign-top fixed-col">';
+echo '</td></tr></table>';
 
 if ( basename( $_SERVER['PHP_SELF'] ) != 'index.php' )
 {
+	echo '<hr /><table class="width-100p valign-top fixed-col">';
+
 	echo '<tr class="st"><td>';
 
 	$profile_options = array(
@@ -392,52 +392,23 @@ if ( basename( $_SERVER['PHP_SELF'] ) != 'index.php' )
 		//echo SelectInput($staff['SCHOOL_ID'],'staff[SCHOOL_ID]','School',$options,'All Schools');
 	}
 
-	echo '</td></tr>';
+	echo '</td></tr></table>';
 }
-
-echo '<tr class="st"><td>';
-// FJ Moodle integrator: email required
-//echo TextInput($staff['EMAIL'],'staff[EMAIL]',_('Email Address'),'size=12 maxlength=100');
-echo TextInput(
-	issetVal( $staff['EMAIL'], '' ),
-	'staff[EMAIL]',
-	_( 'Email Address' ),
-	'type="email" pattern="[^ @]*@[^ @]*" size=12 maxlength=100' .
-	( ! empty( $_REQUEST['moodle_create_user'] ) || ! empty( $old_user_in_moodle ) ? ' required' : '' ),
-	empty( $_REQUEST['moodle_create_user'] )
-);
-
-// FJ create account.
-
-if ( basename( $_SERVER['PHP_SELF'] ) === 'index.php' )
-{
-	echo '</td><td>';
-}
-else
-{
-	echo '</td><td colspan="2">';
-}
-
-echo TextInput(
-	issetVal( $staff['PHONE'], '' ),
-	'staff[PHONE]',
-	_( 'Phone Number' ),
-	'size=12 maxlength=100'
-);
-
-// FJ create account.
-
-if ( basename( $_SERVER['PHP_SELF'] ) === 'index.php' )
-{
-	echo '</td><td>';
-
-	// Add Captcha.
-	echo CaptchaInput( 'captcha' . rand( 100, 9999 ), _( 'Captcha' ) );
-}
-
-echo '</td></tr></table>';
 
 $_REQUEST['category_id'] = '1';
 $separator = '<hr />';
 
 require_once 'modules/Users/includes/Other_Info.inc.php';
+
+// FJ create account.
+if ( basename( $_SERVER['PHP_SELF'] ) === 'index.php' )
+{
+	echo '<hr />';
+
+	echo '<table class="create-account width-100p valign-top fixed-col"><tr class="st"><td colspan="3">';
+
+	// Add Captcha.
+	echo CaptchaInput( 'captcha' . rand( 100, 9999 ), _( 'Captcha' ) );
+
+	echo '</td></tr></table>';
+}