+++ WHATS_NEW.md
@@ -5,6 +5,14 @@
 New in 5.9
 ----------
 
+**Breaking Changes**:
+- Move Email & Phone Staff Fields to custom fields
+- SQL Rename PHONE column to CUSTOM_200000001 in STAFF table
+Please upgrade the following add-ons:
+- [Staff and Parents Import](https://www.rosariosis.org/modules/staff-parents-import/) module to version 1.3
+- [SMS](https://www.rosariosis.org/modules/sms/) + Premium modules to version 1.5
+- [Public Pages Premium](https://www.rosariosis.org/plugins/public-pages/) plugin to version 1.5
+
 Moodle plugin
 - Import Moodle Users